Allow non-moderated sessions during outage
In the event that a SessionTracker wasn't able to be persisted in the
backend the session was aborted and the user received an error about
being unable to upsert session trackers. While this is required in the
event that the session being established is a moderated session, other
sessions should be allowed to be created without a SessionTracker
resource existing in the backend.

To allow non-moderated sessions to be established in the event there
was a failure persisting the resource we can proceed in the same
manner as when the recording mode does not match the component.
This ensures that a local SessionTracker resource is created,
but any operations performed on it are not attempted to be sent to
the Auth server.

The session `log` was also updated to include a `session_id` field
so that it doesn't need to be added to every individual log message.

Closes #17024 and #17026