Deflake TestProxyWatcher
Changes the ProxyWatcher to only emit an empty proxy list if the
watcher has already been initialized. This prevents an empty
set of proxies from being emitted after the creation of the watcher,
but allows an empty list to be emitted if the proxies become zero
during the life of the watcher.

Closes #18311