[v11] fix node selection (#18340)
The Drone node selector was improperly specified as nodes instead of node, causing Drone to stall builds.

Also removes the concurrency limit on Windows push builds. As there are separate builders for v10, v11 and v12. Given that the concurrency limits are enforced across all builds with the same name, they serialise builds from different teleport versions. it seems wasteful to have a builder for v10 sit idle while waiting a v11 build completes.

The concurrent limits have been left for tag builds, as there are flow-on concurrency concern due to artifact uploading, etc.

Also updates the version of Visual Studio that npm looks for to 2022 in order to match buildbox