Deflake TestResizeTerminal
This test made separate reads, each waiting for some conditions
to be met. Due to timing, it was possible that the events expected
in the second read got included in the first batch of reads and
discarded.

Fix this by reading in a single batch and waiting for all required
conditions to be met.

Closes #13607