Improve site and trusted cluster logging and availability
Moves `UpdateTrustedCluster` logging from debug to info so default
logging level includes when admin operations are performed to
establish or remove trust. Alters `remoteSite` such that it logs
in the same manner as `localSite`

Cherry-picks some of the availability changes made in #18050 to
ensure that agents spawned for trusted clusters are more robust to
connection issues.