Fix CodeQL Integer Conversion Warnings
These warnings are fairly pedantic in most cases, but it was easiest to just fix them as I was reviewing them.
I added to `api/utils/conv,go` the boundary checks for doing this casting.  In many cases this allowed a drop in replacement of one of these functions.

Example CodeQL findings fixed by this:
https://github.com/gravitational/teleport/security/code-scanning/81
https://github.com/gravitational/teleport/security/code-scanning/80
https://github.com/gravitational/teleport/security/code-scanning/79
https://github.com/gravitational/teleport/security/code-scanning/78
https://github.com/gravitational/teleport/security/code-scanning/77
https://github.com/gravitational/teleport/security/code-scanning/70
https://github.com/gravitational/teleport/security/code-scanning/69