pr: Use utils.ContainsExpansion, handle $$, simplify tests
Address PR comments, using `utils.ContainsExpansion` instead of
`strings.ContainsRune`. This means we need to handle `$$` as `$` since
that `$$` is false for `ContainsExpansion` but we cannot check that
literal role name without replacing `$$` with `$`.

Simplify the tests and check that the specific error is being returned
by taking advantage of `errors.Is(nil, nil)` returning `true`.