Add tests for teleterm.Serve with TCP address (#18144)
* Refactor error handling for GetConfigForClient

If GetConfigForClient returns an error, the error is not visible in any
logs or by the client making the request.

Instead of failing, we return a config that doesn't let any client
through. It has ClientAuth set to RequireAndVerifyClientCert, but the
config lacks ClientCAs to verify the client cert against.

This also means that when a client with an invalid cert dials the server,
it's going to fail on net.Conn.Read and not tls.Dialer.DialContext. This
will help us add uniform tests in the next commit.

This use of GetConfigForClient is more similar to what we do in other
parts of the codebase, for example in lib/srv/db/proxyserver.go.

* Add tests for teleterm.Serve with TCP address

Previously we'd only test teleterm.Serve with a unix socket, meaning our
whole TCP setup would not be tested.

Testing the TCP server means that we need to set up proper TLS configs
for clients in tests.