Improve trusted cluster observability (#18484) (#18610)
Moves `UpdateTrustedCluster` logging from debug to info so default
logging level includes when admin operations are performed to
establish or remove trust. Alters `remoteSite` such that it logs
in the same manner as `localSite`

Cherry-picks some of the availability changes made in #18050 to
ensure that agents spawned for trusted clusters are more robust to
connection issues.

* Ensure metric `remote_cluster` reflects current state

The metric wasn't properly updated when remote sites went offline
of when remote cluster resources were removed. Any change to the
remoteSite state or the remoteCluster resource are now accurately
reflected in the metric.

* Add tracking of outbound connections to remote clusters

The metric `trust_clusters` existed and was exported, but was never
used anywhere. Now when the `RemoteClusterTunnelManager` starts
and stops agent pools it will create and delete a counter for the
cluster. Within the `AgentPool` the metric is set to the number
of connected proxies within `updateConnectedProxies`.