Add gatewaytest.MustGenAndSaveCert
This function builds on top of alpnproxytest.MustGenCertSignedWithCA.

In alpnproxy, LocalProxy operates on certs solely in memory through
NewLocalProxy and LocalProxy.SetCerts.

Connect on the other hand assumes that the key pair can be loaded from disk,
so we have to provide a function which generates the certs and then saves
them to file.

In the next commit, we're going to check the subject of the cert, so Connect
tests need a way of generating valid certs.