Take cloud labels into account for application access permissions calculation. (#18681)
The app access service now takes cloud labels into account for determining
permissions to an application. Previously, the app service didn't recognize
cloud labels w/r/t determining whether users had access to a particular
application. This was particularly confusing because this worked in the UI,
i.e. the application would show in the list of applications the user had
permissions to, but the app service itself would reject them. This is because
the dynamic and cloud labels were not being properly injected into the
application before running authorize within the app service.