Do not run parker process for all SSH sessions (#18604)
Parker process is only needed when a user is auto-provision by Teleport. Currently, this process always starts, which causes some problems when SELinux is enabled and may also cause issues with a similar mechanism as AppArmor.
This PR runs only the parker process only when it's really needed not for every SSH session reducing the problem mentioned above.