[v9] Improve `tsh play` JSON & YAML output (#18827)
* tsh: clean up onPlay

The code for `tsh play` is a bit hard to follow
(it's a triple-nested switch statement) because
the command is overloaded to do lots of different
things.

Try to make things clearer by adding docs and breaking
the code up into separate functions.

* Improve `tsh play` JSON & YAML output (#18768)

Before this commit, events in the output of `tsh play --format=yaml` were
separated by a newline. This commit ensures that these events are now
separated by `---`, allowing the command output to play well with `yq`.

For consistency, the output of `tsh play --format=yaml <session-id>`
now mirrors what we get e.g. with `tctl nodes ls --format=yaml`.
The same applies to `tsh play --format json`.

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>