auth: Add SAML web handler (#606)
Add the SAML web handler from the OSS repository. This is a copy and
paste of the `auth.validateSAMLResponse` method with package name
tweaks.

Update the auth plugin to register the auth web handler. This includes
a check if the handler is already registered as the OSS repo still has the
handler. Once the corresponding PR in the OSS repo that removes this
registration has been removed, a follow-up commit can remove this check
here.

Source-commit: gravitational/teleport.e@4ac202b0531085d65d96a8040ccb93d3ca27dab3
Source-pull-requests: gravitational/teleport.e#606