auth: Add OIDC connector (#630)
Add the OIDC SSO connector from the OSS repository, along with its
tests. Update references in e/lib/web to point to e/lib/auth now that
some symbols have moved.

The auth web handler has some temporary setup code so that it can
coexist with the OSS repository which still installs the http handler in
the router. Once the OIDC handler has been removed from the OSS repo,
this setup code will be removed.

Remove the `Log` field from `Plugin` and `SAMLAuthService`. Nothing
actually parameterises this, and it makes it more complicated to log
from standalone functions (vs methods). It also requires passing around
a logger to new things.

This will be an issue with the OIDC connector which has a bunch of
standalone functions that log.

This follows the pattern as `github.com/gravitational/teleportlib/auth`
which has a package-level logger.

Source-commit: gravitational/teleport.e@4eeb1cd493ec6ca024279c3413a587fd9d00fa14
Source-pull-requests: gravitational/teleport.e#630