Edit the Database Access introduction
Closes #17092

The Database Access introduction lists the databases that Teleport
supports, but also includes a list of links to all of our Database
Access guides.

So we don't have to maintain two separate lists, I have removed the
first list of supported databases.

I have also edited the heading of the section that lists our
database-specific guides so it's clearer that they are about specific
databases.