Remove mentions of "setup" as a verb (#18949)
Closes #9683

To reduce confusion, we should use "set up" as a verb, rather than
"setup", and treat "setup" only as a noun.

Note that one use of the term "setup" as a verb is in the Jira Server
guide, which #17546 intended to remove, but neglected to. This change
removes that instance of "setup" along with the rest of the surrounding
guide.