Add proto.AuthService/ListAppSessions
`proto.AuthService/GetAppSessions` is not paginated and can cause loading
a massive set of resources into memory on the Auth server as well as sending responsed
which exceede the gRPC message limits. A new paginated mechanism is added to prevent
unbound retrieval of AppSessions. The GetAppSessions RPC still exists in the
proto.AuthService, but has been deprecated and marked for remove in v13.

The `(api.Client) GetAppSessions` function has been refactored to call ListAppSessions
instead of GetAppSessions to reduce the impact of consumers of the client, while
ensuring that exceeding the grpc message size limits is no longer possible.