Drop usage events after too many retries
This change adds a retry counter to individual events (5) and drops
them after too many failures. If an event is malformed and can never
be submitted, or if the backend is down for an extended period of
time, events are dropped instead of generating excessive log spam.

Additionally, only failed events are resubmitted rather than entire
batches as some events may have submitted successfully.

Also, a few log messages are revised.