[v11] Improve and unify cache logging (#19252)
* Improve and unify cache logging

Re-init the cache errors now provide the error via `WithError` instead
of passing the stringify error into the log message. This allows for
the stack trace of the log to show up in the logs to help determine
which resource caused the error.

Logging in the various collection resources was all converted to
follow the same pattern to ensure consistency.

The `Logger` field of the cache was also moved to a named field instead
of being an embedded field.