checkpoint commit
The key thing to remember here is that the line

fmt.Printf("oidstruct = %x\n", oidstruct)

Prints out the marshaled bytes exactly as you found them in the example.
No clue why you need the extra ".1" on the asn1OID, but that gets it done...
fishy. Should try it without that to see if that works.