DatabaseService resource: rpc calls
In the context of Discover we must be able to know when a
DatabaseService process is running, in order to give feedback to the
user.

To do so, we are creating a new resource: DatabaseService.

It will be similar to the way DatabaseServers work, using the same
heartbeat mechanism.

To ease the review, we'll only add the RPC calls in this PR.
Following PRs will:
- CRUD management over rpc
- DatabaseService hearthbeat
- webapi endpoint