DatabaseService resource: rpc calls (#19363)
In the context of Discover we must be able to know when a DatabaseService process is running, in order to give feedback to the user.
DatabaseService process is the process that teleport binary manages when the `teleport.yaml` has
```yaml
db_service:
  enabled: "yes"
```

To do so, we are creating a new resource: DatabaseService.

It will be similar to the way DatabaseServers work, using the same heartbeat mechanism.

To ease the review, we'll only add the RPC calls in this PR. Following PRs will:
- CRUD management over rpc
- DatabaseService hearthbeat
- webapi endpoint

Part of #19032