Add tctl get handlers.
As we expand functionality into enterprise Teleport, there will be enterprise
only clients that OSS tctl will be unable to access. To support `tctl get` for
objects associated with enterprise only features, GetHandlers have been added.
These are similar to CreateHandlers in that the enterprise `tctl` will be able
to add to these handlers on initialization, allowing us to expand the objects
and types that `tctl` is able to retrieve.