DatabaseService: fix ResourceMatcher field (#19475)
In a previous PR (#19363) we created a new resource type:
DatabaseService

Its only spec property for now is the ResourceMatcher field
This field should mirror what we offer in the configuration within
`db_service.resurces` from `teleport.yaml`.

Even though, the current implementation of
`types.DatabaseService.ResourceMatchers` is convertible to/from
`services.ResourceMatchers` - because its only field (from the latter)
is a list of labels - we would incur in breaking changes later on if the
`services.ResourceMatchers` got new fields.

This new resource is only in master (not yet released) and the backport
to v11 must include this change to prevent a breaking change from
happening in the future.