Disables kubeconfig update on `tsh login` when `--kube-cluster` is empty
This PR disables the update of kubecpmfog when `tsh login` does not have the `--kube-cluster` flag.

The new behavior will update the kubeconfig only if the user provided a
non empty value in  `--kube-cluster`.

This PR also introduces a mechanism to restore the current context on
`tsh logout`. On `tsh login --kube-cluster=<cluster>` or `tsh kube login
<cluster>`, Teleport updates the selected context - if it's not from
Teleport - to add an extension whose key is : `selected_context`. Once
`tsh logout` is called and the active context was a Teleport cluster, we
restore the current context to the cluster that has the
`selected_context` key in extensions.

Fixes #9718