Fix RBAC rules for desktop session recordings (#19629)
Desktop session playback is currently the only playback that leverages
the StreamSessionEvents API (though that will change with RFD 91).

For this API, we were checking for VerbList instead of VerbRead.
(The SSH session APIs were correctly checking VerbRead).

Since all uses of actionForKindSession now use the same verb, I've
removed the verb as an argument to prevent this mistake from happening
again.