Batch Dependabot PRs
Leverages https://github.com/Legal-and-General/dependabot-batcher
to create and/or maintain a merged PR (and branch) of all the
Dependabot PRs and close them as they are merged in, leaving a
single Dependabot PR behind.

The Dependabot schedule is set for Sunday at 9:00am and the batcher
workflow is configured to run Sunday at 9:00pm.

All workflows that run on pull requests open have been configured
to ignore branches prefixed with `dependabot/`. Since the dependabot PRs
are all going to be closed and batched there is no need to waste CI
minutes on them. Any new workflows will require the same check to
prevent running on dependabot PRs.