[v11] backport #19695 and #19732 (#19763)
* Update Kubernetes Dependencies (#19695)

This PR replaces the following PRs opened by dependabot:
- #19678
- #19677

It also bumps:

- Bump k8s.io/api from v0.25.4 to v0.26.0
- Bump k8s.io/apiextensions-apiserver from v0.25.4 to v0.26.0
- Bump k8s.io/apimachinery from v0.25.4 to v0.26.0
- Bump k8s.io/apiserver from v0.25.4 to v0.26.0
- Bump k8s.io/cli-runtime from v0.25.4 to v0.26.0
- Bump k8s.io/kubectl from v0.25.4 to v0.26.0
- Bump k8s.io/utils v0.0.0-20220728103510-ee6ede2d64ed to v0.0.0-20221128185143-99ec85e7a448
- Bump sigs.k8s.io/controller-runtime from v0.13.1 to v0.14.1

* Fix goroutine leak when using  `multiResizeQueue` and `spdy` connection leak (#19732)

This PR fixes multiple goroutine and memory leaks when interactive sessions are used. 

- When the session terminates, the `multiResizeQueue` never returns, and the resize stream goroutine blocks. 
- A goroutine leak exited when the server received resizing events after the connection terminated - this happens with fast exec requests.
- A memory leak existed when users tried to leave the session after the `session.tracker` was closed. 

This PR also releases the connection monitor earlier. When the server is under heavy load, it might take a while for the connection to return an `EOF` - which triggers the service monitor automatic release - and the service monitor resources were leaking until the server resumed normal operation.

It also fixes reloads when new parties join and leave the `multiResizeQueue`.