Include active access requests into Kube Proxy credentials cache key
When users use multiple access requests or try to access a cluster
without the correct access request, Teleport caches the invalid
credentials and does not allow users to access their clusters even when
the user presents a certificate that allows access to the resource.

This situation happens because Kubernetes Proxy stores in a TTL map the
user's certificate to avoid generating and signing it each time the user
hits the proxy. The lookup in cache happens using a key that includes
the `kubeCluster`, `username`, `certificate_expiration`, `kube_users`
and `kube_groups` but didn't include the `active_requests`.

This PR adds the `active_requests` into the cache key to
distinguish different access requests for the same user.

Fixes #19884